+++ max_auc_2/models_test/models_test_STEMI_cutpoints_8.py
@@ -58,7 +58,7 @@ def custom_specificity_score(y_true, proba, threshold=threshold):
 data = Data("STEMI.xlsx", STEMI=True)
 
 predictors = ['Возраст', 'NER1', 'SIRI', 'СОЭ', 'TIMI после', 'СДЛА', 'Killip',
-              'RR 600-1200', 'интервал PQ 120-200', 'NBR1', 'SII']
+              'RR 600-1200', 'интервал PQ 120-200', 'NBR1', 'SII', 'EOS', 'NEUT']
 
 data.prepare(predictors, "isAFAfter", [], scale_data=False)
 
@@ -70,56 +70,80 @@ for i, name in enumerate(predictors):
 df.loc[:, 'isAFAfter'] = data.y[:]
 
 
-"""
-#   Возраст ≥64.6
-#   СДЛА ≥25.1
-df.loc[(df['Возраст'] >= 64.6) & (df['СДЛА'] >= 25.1), ('F1')] = 1
-df.loc[(df['Возраст'] < 64.6) | (df['СДЛА'] < 25.1), ('F1')] = 0
-"""
+# Пороги:
+# Возраст ≥65.93737748427554
+# NER1 ≥2387.6009243842545
+# SIRI ≥5.022349948362129
+# СОЭ ≥29.6049917083199
+# TIMI после ≤1.7075872835465769
+# СДЛА ≥22.230746149212216
+# Killip ≥3.0257566769725166
+# RR 600-1200 ≤543.4873416057574
+# интервал PQ 120-200 ≥218.96049246714693
+# RR 600-1200_ ≥1336.2440263241833
 
-#   Возраст ≥64.8
-#   Killip ≥1.0
-df.loc[(df['Возраст'] >= 64.8), ('F1')] = 1
-df.loc[(df['Возраст'] < 64.8), ('F1')] = 0
+# Фенотипы:
+# Возраст & (СДЛА)
+# NER1 & (СОЭ)
+# SIRI
+# СОЭ & (СДЛА)
+# TIMI после
+# СДЛА & (Возраст | SIRI | Killip | TIMI после)
+# Killip & (СДЛА)
+# RR 600-1200 & (Возраст | Killip)
+# интервал PQ 120-200 & (SIRI | RR 600-1200)
+# RR 600-1200_ & (Возраст)
 
-#   NER1 ≥48.92500000000007
-#   СОЭ ≥11.499999999999998
-df.loc[(df['NER1'] >= 48.925) & (df['СОЭ'] >= 11.499), ('F2')] = 1
-df.loc[(df['NER1'] < 48.925) | (df['СОЭ'] < 11.499), ('F2')] = 0
-
-#   NER1 ≥53.47409090909093
-#   СДЛА ≥25.200000000000003
-df.loc[(df['NER1'] >= 53.474090909) & (df['СДЛА'] >= 25.2), ('F3')] = 1
-df.loc[(df['NER1'] < 53.474090909) | (df['СДЛА'] < 25.2), ('F3')] = 0
-
-#   SIRI ≥2.815959821263025
-#   TIMI после ≤2.0
-df.loc[(df['SIRI'] >= 2.8159) & (df['TIMI после'] <= 2.0), ('F4')] = 1
-df.loc[(df['SIRI'] < 2.8159) | (df['TIMI после'] > 2.0), ('F4')] = 0
-
-#   SIRI ≥4.625897044787077
-df.loc[(df['SIRI'] >= 4.625897), ('F5')] = 1
-df.loc[(df['SIRI'] < 4.625897), ('F5')] = 0
-
-#   СОЭ ≥25.3
-#   СДЛА ≥24.400000000000002
-df.loc[(df['СОЭ'] >= 25.3) & (df['СДЛА'] >= 24.4), ('F6')] = 1
-df.loc[(df['СОЭ'] < 25.3) | (df['СДЛА'] < 24.4), ('F6')] = 0
-
-#   СОЭ ≥28.1
-#   интервал PQ 120-200 ≤120.0
-df.loc[(df['СОЭ'] >= 28.1) & (df['интервал PQ 120-200'] <= 120), ('F7')] = 1
-df.loc[(df['СОЭ'] < 28.1) | (df['интервал PQ 120-200'] > 120), ('F7')] = 0
-
-#   СОЭ ≥9.9
-#   RR 600-1200_ ≤595.0
-df.loc[(df['СОЭ'] >= 9.9) & (df['RR 600-1200'] <= 595.0), ('F8')] = 1
-df.loc[(df['СОЭ'] < 9.9) | (df['RR 600-1200'] > 595.0), ('F8')] = 0
+# Веса: [ 0.34709376 -0.08150446  0.66487236  0.70630475  0.73151799  0.88766336
+#   1.06919782  1.18283922  0.28616948  0.        ]
 
 
 
+# Возраст & (СДЛА)
+df.loc[(df['Возраст'] >= 65.9373) & ((df['СДЛА'] >= 22.2307)), ('F1')] = 1
+df.loc[(df['Возраст'] < 65.9373) | ((df['СДЛА'] < 22.2307)), ('F1')] = 0
 
-features = [ 'F1', 'F2', 'F3', 'F4', 'F5', 'F6', 'F7', 'F8' ]
+# SIRI
+df.loc[(df['SIRI'] >= 5.0223), ('F3')] = 1
+df.loc[(df['SIRI'] < 5.0223), ('F3')] = 0
+
+# СОЭ & (СДЛА)
+df.loc[(df['СОЭ'] >= 29.6049) & ((df['СДЛА'] >= 22.2307)), ('F4')] = 1
+df.loc[(df['СОЭ'] < 29.6049) | ((df['СДЛА'] < 22.2307)), ('F4')] = 0
+
+# TIMI после
+df.loc[(df['TIMI после'] <= 1.707587), ('F5')] = 1
+df.loc[(df['TIMI после'] > 1.707587), ('F5')] = 0
+
+# СДЛА & (Возраст | SIRI | Killip | TIMI после)
+df.loc[(df['СДЛА'] >= 22.2307) & ((df['Возраст'] >= 65.937) |
+                                  (df['SIRI'] >= 5.0223) |
+                                  (df['Killip'] >= 3.02575) |
+                                  (df['TIMI после'] <= 1.707587)), ('F6')] = 1
+df.loc[(df['СДЛА'] < 22.2307) | ((df['Возраст'] < 65.937) &
+                                  (df['SIRI'] < 5.0223) &
+                                  (df['Killip'] < 3.02575) &
+                                  (df['TIMI после'] > 1.707587)), ('F6')] = 0
+
+# Killip & (СДЛА)
+df.loc[(df['Killip'] >= 3.02575) & ((df['СДЛА'] >= 22.2307)), ('F7')] = 1
+df.loc[(df['Killip'] < 3.02575) | ((df['СДЛА'] < 22.2307)), ('F7')] = 0
+
+# RR 600-1200 & (Возраст | Killip)
+df.loc[(df['RR 600-1200'] <= 543.48734) & ((df['Возраст'] >= 65.937) |
+                                           (df['Killip'] >= 3.02575)), ('F8')] = 1
+df.loc[(df['RR 600-1200'] > 543.48734) | ((df['Возраст'] < 65.937) &
+                                           (df['Killip'] < 3.02575)), ('F8')] = 0
+
+# интервал PQ 120-200 & (SIRI | RR 600-1200)
+df.loc[(df['интервал PQ 120-200'] >= 218.9604) & ((df['SIRI'] >= 5.0223) |
+                                                  (df['RR 600-1200'] <= 543.48734)), ('F9')] = 1
+df.loc[(df['интервал PQ 120-200'] < 218.9604) | ((df['SIRI'] < 5.0223) &
+                                                  (df['RR 600-1200'] > 543.48734)), ('F9')] = 0
+
+
+features = [ 'F1', 'F3', 'F4', 'F5', 'F6', 'F7', 'F8', 'F9' ]
+
 
 isModel = 1 # 1 - Logistic
 rm = 100
