+++ max_auc_2/models_test/models_test_STEMI_rect_2.py
@@ -70,111 +70,50 @@ for i, name in enumerate(predictors):
 df.loc[:, 'isAFAfter'] = data.y[:]
 
 
-# Пороги:
-# Возраст ≥67.6141750939809
-# NER1 ≥2851.92356663509
-# SIRI ≥6.675510789705735
-# СОЭ ≥35.46879380317242
-# TIMI после ≤1.8996069134161897
-# СДЛА ≥35.52387717581928
-# Killip ≥3.159671820862726
-# RR 600-1200 ≥870.3493540805769
-# интервал PQ 120-200 ≤197.57468555271487
-# RR 600-1200_ ≤681.4175134782593
+# Возраст ≥59.0
+#   NER1 ≥43.60000000000002
+df.loc[(df['Возраст'] >= 59.0) & (df['NER1'] >= 43.60000000000002), ('F1')] = 1
+df.loc[(df['Возраст'] < 59.0) | (df['NER1'] < 43.60000000000002), ('F1')] = 0
 
-# Фенотипы:
-# Возраст
-# NER1 & (SIRI | СОЭ | TIMI после | СДЛА | Killip | RR 600-1200 | интервал PQ 120-200 | RR 600-1200_)
-# SIRI
-# СОЭ
-# TIMI после
-# СДЛА & (интервал PQ 120-200)
-# Killip
-# RR 600-1200 & (Возраст | NER1 | SIRI | СОЭ | TIMI после | СДЛА | Killip | интервал PQ 120-200 | RR 600-1200_)
-# интервал PQ 120-200 & (Возраст | SIRI | СОЭ | TIMI после | СДЛА | Killip)
-# RR 600-1200_
+#   Возраст ≥64.2
+#   СДЛА ≥25.2
+df.loc[(df['Возраст'] >= 64.2) & (df['СДЛА'] >= 25.2), ('F2')] = 1
+df.loc[(df['Возраст'] < 64.2) | (df['СДЛА'] < 25.2), ('F2')] = 0
 
+#   Возраст ≥57.8
+#   RR 600-1200 ≤597.6
+df.loc[(df['Возраст'] >= 57.8) & (df['RR 600-1200'] <= 597.6), ('F3')] = 1
+df.loc[(df['Возраст'] < 57.8) | (df['RR 600-1200'] > 597.6), ('F3')] = 0
 
-# Возраст
-df.loc[(df['Возраст'] >= 67.61), ('F1')] = 1
-df.loc[(df['Возраст'] < 67.61), ('F1')] = 0
+# NER1 ≥49.10000000000002
+#   СОЭ ≥15.0
+df.loc[(df['NER1'] >= 49.10000000000002) & (df['СОЭ'] >= 15.0), ('F4')] = 1
+df.loc[(df['NER1'] < 49.10000000000002) | (df['СОЭ'] < 15.0), ('F4')] = 0
 
-# NER1 & (SIRI | СОЭ | TIMI после | СДЛА | Killip | RR 600-1200 | интервал PQ 120-200 | RR 600-1200_)
-df.loc[(df['NER1'] >= 2851.92) & ((df['SIRI'] >= 6.675) |
-                                  (df['СОЭ'] >= 35.46) |
-                                  (df['TIMI после'] <= 1.89) |
-                                  (df['СДЛА'] >= 35.52) |
-                                  (df['Killip'] >= 3.15) |
-                                  (df['интервал PQ 120-200'] <= 197.57) |
-                                  (df['RR 600-1200'] <= 681.41)), ('F2')] = 1
-df.loc[(df['NER1'] < 2851.92) | ((df['SIRI'] < 6.675) &
-                                  (df['СОЭ'] < 35.46) &
-                                  (df['TIMI после'] > 1.89) &
-                                  (df['СДЛА'] < 35.52) &
-                                  (df['Killip'] < 3.15) &
-                                  (df['интервал PQ 120-200'] > 197.57) &
-                                  (df['RR 600-1200'] > 681.41)), ('F2')] = 0
+# NER1 ≥51.379999999999995
+#   СДЛА ≥25.0
+df.loc[(df['NER1'] >= 51.379999999999995) & (df['СДЛА'] >= 25.0), ('F5')] = 1
+df.loc[(df['NER1'] < 51.379999999999995) | (df['СДЛА'] < 25.0), ('F5')] = 0
 
-# SIRI
-df.loc[(df['SIRI'] >= 6.675), ('F3')] = 1
-df.loc[(df['SIRI'] < 6.675), ('F3')] = 0
+# SIRI ≥3.597711721733971
+#   TIMI после ≤2.0
+df.loc[(df['SIRI'] >= 3.597711) & (df['TIMI после'] <= 2.0), ('F6')] = 1
+df.loc[(df['SIRI'] < 3.597711) | (df['TIMI после'] > 2.0), ('F6')] = 0
 
-# СОЭ
-df.loc[(df['СОЭ'] >= 35.46), ('F4')] = 1
-df.loc[(df['СОЭ'] < 35.46), ('F4')] = 0
+#  SIRI ≥4.731617878120411
+#   СДЛА ≥22.200000000000003
+df.loc[(df['SIRI'] >= 4.73161) & (df['СДЛА'] >= 22.200000000000003), ('F7')] = 1
+df.loc[(df['SIRI'] < 4.73161) | (df['СДЛА'] < 22.200000000000003), ('F7')] = 0
 
-# TIMI после
-df.loc[(df['TIMI после'] <= 1.89), ('F5')] = 1
-df.loc[(df['TIMI после'] > 1.89), ('F5')] = 0
-
-# СДЛА & (интервал PQ 120-200)
-df.loc[(df['СДЛА'] >= 35.52) & (df['интервал PQ 120-200'] <= 197.57), ('F6')] = 1
-df.loc[(df['СДЛА'] < 35.52) | (df['интервал PQ 120-200'] > 197.57), ('F6')] = 0
-
-# Killip
-df.loc[(df['Killip'] >= 3.15), ('F7')] = 1
-df.loc[(df['Killip'] < 3.15), ('F7')] = 0
-
-# RR 600-1200 & (Возраст | NER1 | SIRI | СОЭ | TIMI после | СДЛА | Killip | интервал PQ 120-200 | RR 600-1200_)
-df.loc[(df['RR 600-1200'] >= 870.34) & ((df['Возраст'] >= 67.61) |
-                                        (df['NER1'] >= 2851.92) |
-                                        (df['SIRI'] >= 6.675) |
-                                        (df['СОЭ'] >= 35.46) |
-                                        (df['TIMI после'] <= 1.89) |
-                                        (df['СДЛА'] >= 35.52) |
-                                        (df['Killip'] >= 3.15) |
-                                        (df['интервал PQ 120-200'] <= 197.57) |
-                                        (df['RR 600-1200'] <= 681.41)), ('F8')] = 1
-df.loc[(df['RR 600-1200'] < 870.34) | ((df['Возраст'] < 67.61) &
-                                        (df['NER1'] < 2851.92) &
-                                        (df['SIRI'] < 6.675) &
-                                        (df['СОЭ'] < 35.46) &
-                                        (df['TIMI после'] > 1.89) &
-                                        (df['СДЛА'] < 35.52) &
-                                        (df['Killip'] < 3.15) &
-                                        (df['интервал PQ 120-200'] > 197.57) &
-                                        (df['RR 600-1200'] > 681.41)), ('F8')] = 0
-
-# интервал PQ 120-200 & (Возраст | SIRI | СОЭ | TIMI после | СДЛА | Killip)
-df.loc[(df['интервал PQ 120-200'] <= 197.57) & ((df['Возраст'] >= 67.61) |
-                                                 (df['SIRI'] >= 6.675) |
-                                                 (df['СОЭ'] >= 35.46) |
-                                                 (df['TIMI после'] <= 1.89) |
-                                                 (df['СДЛА'] >= 35.52) |
-                                                 (df['Killip'] >= 3.15)), ('F9')] = 1
-df.loc[(df['интервал PQ 120-200'] > 197.57) | ((df['Возраст'] < 67.61) &
-                                                 (df['SIRI'] < 6.675) &
-                                                 (df['СОЭ'] < 35.46) &
-                                                 (df['TIMI после'] > 1.89) &
-                                                 (df['СДЛА'] < 35.52) &
-                                                 (df['Killip'] < 3.15)), ('F9')] = 0
-
-# RR 600-1200_
-df.loc[(df['RR 600-1200'] <= 681.41), ('F10')] = 1
-df.loc[(df['RR 600-1200'] > 681.41), ('F10')] = 0
+# RR 600-1200 ≤592.0
+#   интервал PQ 120-200 ≥204.0
+df.loc[(df['RR 600-1200'] <= 592.0) & (df['интервал PQ 120-200'] >= 204.0), ('F8')] = 1
+df.loc[(df['RR 600-1200'] > 592.0) | (df['интервал PQ 120-200'] < 204.0), ('F8')] = 0
 
 
-features = [ 'F1', 'F2', 'F3', 'F4', 'F5', 'F6', 'F7', 'F8', 'F9', 'F10' ]
+
+
+features = [ 'F1', 'F2', 'F3', 'F4', 'F5', 'F6', 'F7', 'F8' ]
 
 isModel = 1 # 1 - Logistic
 rm = 100
